Refactor: Locking Mechanism refactored with spinner

+++ src/Dashboard.jsx
@@ -9,6 +9,7 @@ import PublishedWithChangesIcon from "@mui/icons-material/PublishedWithChanges";
 import DeleteSweepIcon from "@mui/icons-material/DeleteSweep";
 import { useAuth } from "./AuthContext";
 import ErrorDisplay from "./Components/ErrorDisplay";
+import { useUI } from "./UIContext";
 
 const Dashboard = ({
 	dbCards,
@@ -22,6 +23,7 @@ const Dashboard = ({
 }) => {
 	const [localCards, setLocalCards] = useState([]);
 	const { user, userState } = useAuth();
+	const { isEditingLock, editingLockRef } = useUI();
 
 	const cards =
 		userState === "loggedIn"
@@ -51,6 +53,7 @@ const Dashboard = ({
 	}, [dbCards]); // New Use Effect
 
 	async function handleDeleteAll() {
+		if (editingLockRef === true) return;
 		if (userState === "loggedIn") {
 			await deleteAllCardsInDB();
 			readCardsFromDB();
@@ -59,54 +62,13 @@ const Dashboard = ({
 		}
 	}
 	async function handleClearAllDoneTasks() {
+		if (editingLockRef === true) return;
 		const clearedDoneCards = cards.filter((card) => !card.checked);
 		if (userState === "loggedIn") {
 			await clearDoneCardsInDB(clearedDoneCards);
 			readCardsFromDB();
 		} else {
 			setLocalCards(clearedDoneCards);
-		}
-	}
-
-	function handleCheckedChanged(id) {
-		if (userState === "loggedIn") {
-			cards.map((card) => {
-				if (card.id === id) {
-					updateCardsInDB(id, { checked: !card.checked });
-				} else {
-					return card;
-				}
-			});
-		} else {
-			const newCards = cards.map((card) => {
-				if (card.id === id) {
-					return { ...card, checked: !card.checked };
-				} else {
-					return card;
-				}
-			});
-			setLocalCards(newCards);
-		}
-	}
-
-	function handlePriorityChanged(id) {
-		if (userState === "loggedIn") {
-			cards.map((card) => {
-				if (card.id === id) {
-					updateCardsInDB(id, { highPriority: !card.highPriority });
-				} else {
-					return card;
-				}
-			});
-		} else {
-			const newCards = cards.map((card) => {
-				if (card.id === id) {
-					return { ...card, highPriority: !card.highPriority };
-				} else {
-					return card;
-				}
-			});
-			setLocalCards(newCards);
 		}
 	}
 
@@ -143,122 +105,152 @@ const Dashboard = ({
 		}
 	}
 
-	function handleTextChange(updatedText) {
-		cards.map((card) => {
-			if (card.id === selectedCardID) {
-				updateCardsInDB(selectedCardID, { text: updatedText });
-			} else {
-				return card;
-			}
-		});
+	function updateCardById(id, updatedFields) {
+		if (userState === "guest") {
+			console.log("updated fields = ", updatedFields);
+			const updatedCards = localCards.map((card) => {
+				return card.id === id ? { ...card, ...updatedFields } : card;
+			});
+			console.log("updated cards= ", updatedCards);
+			setLocalCards(updatedCards);
+		} else if (userState === "loggedIn") {
+			updateCardsInDB(id, updatedFields);
+		}
 	}
+
+	function handleTextChange(id, updatedText, flagProps = {}) {
+		console.log(
+			"handle text change called in dashboard with flagProps = ",
+			flagProps
+		);
+		updateCardById(id, { text: updatedText, ...flagProps });
+	}
+
+	function handleFlagToggleChange(id, flagName, currentFlagValue, currentText) {
+		updateCardById(id, { [flagName]: !currentFlagValue, text: currentText });
+	}
+
 	return (
 		<div className="main-page-container">
 			<div className="main">
 				<div>
 					<Header />
 				</div>
-				<div className="summary">
-					<div id="summary-heading">
-						<h2 className="text-2xl font-semibold my-2">Summary</h2>
+				<div
+					className={`dashboard-content ${
+						isEditingLock ? "dashboard-is-locked" : ""
+					}`}
+				>
+					<button>Test</button>
+					<div className="summary">
+						<div id="summary-heading">
+							<h2 className="text-2xl font-semibold my-2">Summary</h2>
+						</div>
+						<h3 className="text-xl font-medium my-1"> You Have:</h3>
+						<p className="text-gray-700">{cards.length} Total Tasks</p>
+						<p className="text-gray-700">
+							{highPriorityCardsTotal} High Priority Tasks
+						</p>
+						<p className="text-gray-700">{doneCardsTotal} Tasks Done</p>
+
+						<div>
+							<Tooltip title="Clear Done Tasks" placement="left">
+								<IconButton
+									disabled={editingLockRef.current}
+									onClick={handleClearAllDoneTasks}
+								>
+									<PublishedWithChangesIcon fontSize="large" color="primary" />
+								</IconButton>
+							</Tooltip>
+
+							<Tooltip title="Delete All Tasks" placement="right">
+								<IconButton
+									disabled={editingLockRef.current}
+									onClick={handleDeleteAll}
+								>
+									<DeleteSweepIcon fontSize="large" color="primary" />
+								</IconButton>
+							</Tooltip>
+						</div>
 					</div>
-					<h3 className="text-xl font-medium my-1"> You Have:</h3>
-					<p className="text-gray-700">{cards.length} Total Tasks</p>
-					<p className="text-gray-700">
-						{highPriorityCardsTotal} High Priority Tasks
-					</p>
-					<p className="text-gray-700">{doneCardsTotal} Tasks Done</p>
-
-					<div>
-						<Tooltip title="Clear Done Tasks" placement="left">
-							<IconButton onClick={handleClearAllDoneTasks}>
-								<PublishedWithChangesIcon fontSize="large" color="primary" />
-							</IconButton>
-						</Tooltip>
-
-						<Tooltip title="Delete All Tasks" placement="right">
-							<IconButton onClick={handleDeleteAll}>
-								<DeleteSweepIcon fontSize="large" color="primary" />
-							</IconButton>
-						</Tooltip>
+					<div className="draft-card-container">
+						<DraftCard
+							onAdd={addCard}
+							isAdding={isAdding}
+							disabled={editingLockRef.current}
+						/>
 					</div>
+
+					{!highPriorityHidden && (
+						<>
+							<div className="swimlane-heading" id="high-priority-heading">
+								<h3>High Priority</h3>
+							</div>
+
+							<div className="high-priority-cards-container">
+								{highPriorityCards.map((card) => (
+									<Card
+										key={card.renderKey}
+										id={card.id}
+										text={card.text}
+										checked={card.checked}
+										highPriority={card.highPriority}
+										onDelete={deleteCard}
+										onTextUpdate={handleTextChange}
+										onSelect={selectCard}
+										onFlagToggle={handleFlagToggleChange}
+									/>
+								))}
+							</div>
+						</>
+					)}
+
+					{!allOtherCardsHidden && (
+						<>
+							<div className="swimlane-heading" id="all-other-tasks-heading">
+								<h3>All Other Tasks</h3>
+							</div>
+							<div className="cards-container">
+								{allOtherCards.map((card) => (
+									<Card
+										key={card.renderKey}
+										id={card.id}
+										text={card.text}
+										checked={card.checked}
+										highPriority={card.highPriority}
+										onDelete={deleteCard}
+										onTextUpdate={handleTextChange}
+										onSelect={selectCard}
+										onFlagToggle={handleFlagToggleChange}
+									/>
+								))}
+							</div>
+						</>
+					)}
+
+					{!doneCardsHidden && (
+						<>
+							<div className="swimlane-heading" id="done-tasks-heading">
+								<h3>Tasks Done</h3>
+							</div>
+							<div className="done-cards-container">
+								{doneCards.map((card) => (
+									<Card
+										key={card.renderKey}
+										id={card.id}
+										text={card.text}
+										checked={card.checked}
+										highPriority={card.highPriority}
+										onDelete={deleteCard}
+										onTextUpdate={handleTextChange}
+										onSelect={selectCard}
+										onFlagToggle={handleFlagToggleChange}
+									/>
+								))}
+							</div>
+						</>
+					)}
 				</div>
-				<div className="draft-card-container">
-					<DraftCard onAdd={addCard} isAdding={isAdding} />
-				</div>
-
-				{!highPriorityHidden && (
-					<>
-						<div className="swimlane-heading" id="high-priority-heading">
-							<h3>High Priority</h3>
-						</div>
-
-						<div className="high-priority-cards-container">
-							{highPriorityCards.map((card) => (
-								<Card
-									key={card.renderKey}
-									id={card.id}
-									text={card.text}
-									checked={card.checked}
-									highPriority={card.highPriority}
-									onCheckedChange={handleCheckedChanged}
-									onPriorityChange={handlePriorityChanged}
-									onDelete={deleteCard}
-									onTextUpdate={handleTextChange}
-									onSelect={selectCard}
-								/>
-							))}
-						</div>
-					</>
-				)}
-
-				{!allOtherCardsHidden && (
-					<>
-						<div className="swimlane-heading" id="all-other-tasks-heading">
-							<h3>All Other Tasks</h3>
-						</div>
-						<div className="cards-container">
-							{allOtherCards.map((card) => (
-								<Card
-									key={card.renderKey}
-									id={card.id}
-									text={card.text}
-									checked={card.checked}
-									highPriority={card.highPriority}
-									onCheckedChange={handleCheckedChanged}
-									onPriorityChange={handlePriorityChanged}
-									onDelete={deleteCard}
-									onTextUpdate={handleTextChange}
-									onSelect={selectCard}
-								/>
-							))}
-						</div>
-					</>
-				)}
-
-				{!doneCardsHidden && (
-					<>
-						<div className="swimlane-heading" id="done-tasks-heading">
-							<h3>Tasks Done</h3>
-						</div>
-						<div className="done-cards-container">
-							{doneCards.map((card) => (
-								<Card
-									key={card.renderKey}
-									id={card.id}
-									text={card.text}
-									checked={card.checked}
-									highPriority={card.highPriority}
-									onCheckedChange={handleCheckedChanged}
-									onPriorityChange={handlePriorityChanged}
-									onDelete={deleteCard}
-									onTextUpdate={handleTextChange}
-									onSelect={selectCard}
-								/>
-							))}
-						</div>
-					</>
-				)}
 			</div>
 
 			<div>
